% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.iupac_het}
\alias{.iupac_het}
\title{IUPAC heterozygote code for an allele pair (internal HapMap helper)}
\usage{
.iupac_het(a0, a1)
}
\description{
IUPAC heterozygote code for an allele pair (internal HapMap helper)
}
\keyword{internal}
